PDFBOX-5660: remove redundant cast

+++ src/main/java/org/apache/pdfbox/jbig2/segments/Table.java
@@ -63,7 +63,7 @@ public class Table implements SegmentData
         htPS = (int) ((subInputStream.readBits(3) + 1) & 0xf);
 
         /* Bit 0 */
-        htOutOfBand = (int) subInputStream.readBit();
+        htOutOfBand = subInputStream.readBit();
 
         htLow = (int) subInputStream.readBits(32); // & 0xffffffff);
         htHigh = (int) subInputStream.readBits(32); // & 0xffffffff);
